pyproject.toml
[project]
name = "aidocument"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aspose-words>=25.10.0",
    "beautifulsoup4>=4.14.2",
    "docx2pdf>=0.1.8",
    "fastapi>=0.119.0",
    "markdown>=3.9",
    "openai>=2.5.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfkit>=1.0.0",
    "pdfminer-six>=20250506",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = []
